Move around modules
Move the command object modules over to the eponymous command interface
module.  Move the command object out of the command module and into the
command object module.
