+++ libqtile/layout/base.py
@@ -21,13 +21,13 @@
 
 import copy
 from abc import ABCMeta, abstractmethod
-
-from .. import command, configurable
-
 from typing import Any, List, Tuple  # noqa: F401
 
+from .. import configurable
+from libqtile.command_object import CommandObject
 
-class Layout(command.CommandObject, configurable.Configurable, metaclass=ABCMeta):
+
+class Layout(CommandObject, configurable.Configurable, metaclass=ABCMeta):
     """This class defines the API that should be exposed by all layouts"""
     @classmethod
     def _name(cls):
@@ -47,7 +47,7 @@ class Layout(command.CommandObject, configurable.Configurable, metaclass=ABCMeta
         if "name" not in config:
             config["name"] = self.__class__.__name__.lower()
 
-        command.CommandObject.__init__(self)
+        CommandObject.__init__(self)
         configurable.Configurable.__init__(self, **config)
         self.add_defaults(Layout.defaults)
